Vectors of vectors just became a mess in julia
The code was based on ruby code. For ruby or python it would be natural to create vectors of vectors, but I realised this is not how you are supposed to work with Julia. writecsv can not deal with vectors of vectors properly.

In my experience vectors of vectors just proved awkward and confusing to use in Julia. Everything went a lot smoother once I started using 2D arrays resized to the correct size before filling it up. Something to keep in mind.