Plot curves for demand, supply and price
The curves are not completely similar to the ones in the book but the trends seem to be the same. Possible explanation is that we start indices with 1 rather than 0 so some cycles might be different. Most notably there are no early swings in prices while it is above 5 dollars. My simulation seem to move much faster towards 5 dollars in price than book example. A possible explanation is that supply is not constantly below demand as seen in the book. It seems natural that supply should be below demand, so this needs to be investigated further.